Original commit
The original commit includes v1.0 that was laid out for common anode but
the display we received was common cathode. Version 1.1 of the PCB
layout works with common cathode.

The code is very preliminary. We can get data back and forth using I2C
pretty well. Eventually I need to get this connected with the updated
version of SevSeg.